+++ src/views/Application.vue
@@ -2,12 +2,17 @@
   <div id="application">
     <ec-application
       :application.sync="application"
+      :is-settings-visible="isSettingsVisible"
+      :settings-open-collapse="settingsOpenCollapse"
       @click:back="goToMarket"
       @install="onInstall"
       @uninstall="goToMarket"
     >
       <template #settings>
         <slot name="settings"/>
+      </template>
+      <template #settings-prepend>
+        <slot name="settings-prepend"/>
       </template>
       <template #settings-append>
         <slot name="settings-append"/>